requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
googlemaps==4.10.0
idna==3.7
joblib==1.4.2
numpy==2.0.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
threadpoolctl==3.5.0
urllib3==2.2.2
